requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pypdf==5.4.0
python-dotenv==1.0.1
pytest==8.3.5
